+++ README.md
@@ -91,7 +91,7 @@ const server = fa({
 const { feed } = server
 
 feed.on('ready', () => {
-  server.listen(3000)
+  server.listen(3000, '0.0.0.0')
 })
 ```
