+++ bin/index.js
@@ -5,6 +5,7 @@ const args = require('args')
 const nodeVersion = require('node-version')
 const updateNotifier = require('update-notifier')
 const path = require('path')
+const fa = require('../lib')
 
 if (nodeVersion.major < 6) {
   console.error(
@@ -31,8 +32,13 @@ args
   .option(['m', 'memory'], 'Use in-memory storage', false, Boolean)
 
 const flags = args.parse(process.argv, { name: pkg.name })
-require('../lib')(flags, () => {
-  console.log(
-    '⚡ fair-analytics listening on ' + flags.host + ':' + flags.port + ' ⚡'
-  )
+const server = fa(flags)
+const { feed } = server
+
+feed.on('ready', () => {
+  server.listen(flags.port, flags.host, () => {
+    console.log(
+      '⚡ fair-analytics listening on ' + flags.host + ':' + flags.port + ' ⚡'
+    )
+  })
 })
