+++ lib/index.js
@@ -1,11 +1,16 @@
-const fa = require('./fair-analytics')
+const createFeed = require('./feed')
+const createServer = require('./server')
+const shareFeed = require('./share')
+const createSubscriptions = require('./wire-subscriptions')
+const sse = require('./sse')
+const createStatsDb = require('./stats-db')
 
-module.exports = function run (flags, cb) {
-  const server = fa(flags)
-  const { feed } = server
-
-  feed.on('ready', () => {
-    server.listen(flags.port)
-    cb && cb()
-  })
+module.exports = function fa (flags) {
+  const statsDb = createStatsDb(flags.storageDirectory, flags.memory)
+  const broadcast = createSubscriptions(sse, statsDb)
+  const feed = createFeed(flags.storageDirectory, flags.memory)
+  feed.on('ready', shareFeed.bind(null, feed))
+  const server = createServer(feed, broadcast, sse, statsDb, flags.origin)
+  server.feed = feed
+  return server
 }
