+++ package.json
@@ -2,7 +2,7 @@
   "name": "fair-analytics",
   "version": "1.1.0",
   "description": "Fair Analytics",
-  "main": "fair-analytics.js",
+  "main": "lib/index.js",
   "engines": {
     "node": ">=6"
   },
